use flow check and move flowtype-plugin to top level
